moved some things around to clean up project
