+++ clifscatalogv2/app/functional_components/favorite/Favorite.tsx
@@ -1,5 +1,4 @@
 "use client";
-
 import { Session } from "next-auth";
 
 interface Props {
